Support url encode/decode function
